chore(wizard-refactor): extract tools/update_setup (~370 lines)
Largest non-elephant tool. Pinned by 3 dispatcher tests (REJECTS /
ALLOWS / auto-corrects progressionMode) + 5 partial-save vitests in
lib/chat/__tests__/wizard-tool-executor-partial-save.test.ts. 61/61 green
across all 6 wizard test files.

Drops the monolith's resolveCourseByName / resolveSubjectByName /
resolveInstitutionByName imports — they were only used by update_setup
and ensure-institution-and-domain (which re-imports them).
